Refactor PromQL classic histogram handling
Extract shared classic histogram state and planning hooks to avoid
duplicating quantile infrastructure in additional histogram functions.
