+++ x-pack/plugin/esql/compute/src/test/java/org/elasticsearch/compute/aggregation/PromqlHistogramQuantileGroupingAggregatorFunctionTests.java
@@ -7,7 +7,7 @@
 
 package org.elasticsearch.compute.aggregation;
 
-import org.elasticsearch.compute.aggregation.PromqlHistogramQuantileStates.Bucket;
+import org.elasticsearch.compute.aggregation.PromqlHistogramStates.Bucket;
 import org.elasticsearch.compute.data.Block;
 import org.elasticsearch.compute.data.BlockFactory;
 import org.elasticsearch.compute.data.DoubleBlock;
