+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/plan/logical/promql/HistogramFunctionCall.java
@@ -1,0 +1,69 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+package org.elasticsearch.xpack.esql.plan.logical.promql;
+
+import org.elasticsearch.xpack.esql.core.expression.Attribute;
+import org.elasticsearch.xpack.esql.core.expression.Expression;
+import org.elasticsearch.xpack.esql.core.expression.FieldAttribute;
+import org.elasticsearch.xpack.esql.core.expression.MetadataAttribute;
+import org.elasticsearch.xpack.esql.core.tree.Source;
+import org.elasticsearch.xpack.esql.expression.promql.function.FunctionType;
+import org.elasticsearch.xpack.esql.expression.promql.function.PromqlFunctionDefinition;
+import org.elasticsearch.xpack.esql.plan.logical.LogicalPlan;
+
+import java.util.List;
+
+import static org.elasticsearch.xpack.esql.plan.logical.promql.PromqlLabels.PROMETHEUS_LABELS_PREFIX;
+
+/**
+ * Base class for PromQL histogram functions that evaluate classic histogram buckets grouped by their {@code le} label.
+ */
+public abstract sealed class HistogramFunctionCall extends PromqlFunctionCall permits HistogramQuantile {
+    public static final String LE_LABEL = "le";
+
+    private List<Attribute> output;
+
+    protected HistogramFunctionCall(Source source, LogicalPlan child, PromqlFunctionDefinition definition, List<Expression> parameters) {
+        super(source, child, definition, parameters);
+    }
+
+    /**
+     * Builds the aggregate expression for this function call:
+     * The aggregate expression will be invoked with the bucket counts and their upper bounds (the le labels).
+     */
+    public abstract Expression buildAggregateFunction(Expression count, Expression upperBound);
+
+    @Override
+    public final List<Attribute> output() {
+        if (output == null) {
+            output = child().output()
+                .stream()
+                .filter(attr -> MetadataAttribute.isTimeSeriesAttributeName(attr.name()) || LE_LABEL.equals(labelName(attr)) == false)
+                .toList();
+        }
+        return output;
+    }
+
+    @Override
+    public final FunctionType functionType() {
+        return FunctionType.HISTOGRAM;
+    }
+
+    private static String labelName(Attribute attribute) {
+        String fieldName;
+        if (attribute instanceof FieldAttribute fieldAttribute) {
+            fieldName = fieldAttribute.fieldName().string();
+        } else {
+            fieldName = attribute.name();
+        }
+        if (fieldName.startsWith(PROMETHEUS_LABELS_PREFIX)) {
+            return fieldName.substring(PROMETHEUS_LABELS_PREFIX.length());
+        }
+        return fieldName;
+    }
+}
